requirements.txt
fastapi
uvicorn
Pillow
accelerate
huggingface_hub
requests
pydantic
colpali-engine
stamina
tqdm
ollama
requests
pdf2image
PyMuPDF
qdrant-client
python-multipart